Changed behavior of UserAgentSearch renderer configuration option. Fixed all .conf files accordingly, uncommented the option for backwards compatibility.
The default value for UserAgentSearch was changed from
"^DefaultUserAgent" to an empty string "".

This means that making the option empty does not accidentally match
every User-Agent.
It also means that it is no longer necessary to uncomment
UserAgentSearch or give it a strange value like
"DontMatchAnything-UsingAdditionalHeaderSearchInstead". You can now
simply empty the value and it will work as expected (nothing is
matched, instead of everything).

If all User-Agents should be matched, use the wildcard character "."
instead.